+++ src/Export/ExportProducts.php
@@ -34,15 +34,30 @@ class ExportProducts implements ExportInterface
         }
     }
 
+    public function getBatchByContext(SalesChannelContext $context, int $limit, int $offset): iterable
+    {
+        $criteria = $this->getCriteria($limit, $offset);
+        $products = $this->productRepository->search($criteria, $context);
+
+        foreach ($products->getElements() as $product) {
+            yield $product;
+        }
+    }
+
     public function getProducedExportEntityType(): string
     {
         return ExportProductEntity::class;
     }
 
-    private function getCriteria(int $batchSize): Criteria
+    private function getCriteria(int $batchSize, ?int $offset = null): Criteria
     {
         $criteria = new Criteria();
         $criteria->setLimit($batchSize);
+
+        if ($offset) {
+            $criteria->setOffset($offset);
+        }
+
         $criteria->addAssociation('categories');
         $criteria->addAssociation('categoriesRo');
         $criteria->addAssociation('children.options.group');
@@ -53,9 +68,11 @@ class ExportProducts implements ExportInterface
         $criteria->addAssociation('seoUrls');
         $criteria->addAssociation('media');
         $criteria->addAssociation('children.cover.media');
+
         foreach ($this->customAssociations as $association) {
             $criteria->addAssociation($association);
         }
+
         $criteria->addFilter(new EqualsFilter('parentId', null));
 
         return $criteria;